backend/requirements.txt
fastapi[standard]==0.124.0
pydantic==2.12.0
anthropic==0.18.1
httpx==0.27.0
python-dotenv==1.0.0
